Add files via upload
A quick and dirty line graph data generator I will use as a data source for other parts of the project. It's basically a simple stand in for the price data I would pull from a crypto exchange. It doesn't create realistic data, but it doesn't need to for what I am using it for.